move banner and prompt hooks to pwa feature
